+++ packages/make-styles/src/shorthands/borderColor.ts
@@ -3,19 +3,24 @@ import type { BorderColorProperty } from 'csstype';
 import type { MakeStylesStrictCSSObject } from '../types';
 import { generateStyles } from './generateStyles';
 
-export function borderColor(all: BorderColorProperty): MakeStylesStrictCSSObject;
-export function borderColor(vertical: BorderColorProperty, horizontal: BorderColorProperty): MakeStylesStrictCSSObject;
+type BorderColorStyle = Pick<
+  MakeStylesStrictCSSObject,
+  'borderTopColor' | 'borderRightColor' | 'borderBottomColor' | 'borderLeftColor'
+>;
+
+export function borderColor(all: BorderColorProperty): BorderColorStyle;
+export function borderColor(vertical: BorderColorProperty, horizontal: BorderColorProperty): BorderColorStyle;
 export function borderColor(
   top: BorderColorProperty,
   horizontal: BorderColorProperty,
   bottom: BorderColorProperty,
-): MakeStylesStrictCSSObject;
+): BorderColorStyle;
 export function borderColor(
   top: BorderColorProperty,
   right: BorderColorProperty,
   bottom: BorderColorProperty,
   left: BorderColorProperty,
-): MakeStylesStrictCSSObject;
+): BorderColorStyle;
 
 /**
  * A function that implements CSS spec conformant expansion for "borderColor"
@@ -28,6 +33,6 @@ export function borderColor(
  *
  * See https://developer.mozilla.org/en-US/docs/Web/CSS/border-color
  */
-export function borderColor(...values: BorderColorProperty[]): MakeStylesStrictCSSObject {
+export function borderColor(...values: BorderColorProperty[]): BorderColorStyle {
   return generateStyles('border', 'Color', ...values);
 }