+++ packages/make-styles/src/shorthands/borderTop.ts
@@ -1,16 +1,15 @@
 import type { BorderColorProperty, BorderStyleProperty, BorderWidthProperty } from 'csstype';
 import type { MakeStylesStrictCSSObject, MakeStylesCSSValue } from '../types';
 
-export function borderTop(width: BorderWidthProperty<MakeStylesCSSValue>): MakeStylesStrictCSSObject;
-export function borderTop(
-  width: BorderWidthProperty<MakeStylesCSSValue>,
-  style: BorderStyleProperty,
-): MakeStylesStrictCSSObject;
+type BorderTopStyle = Pick<MakeStylesStrictCSSObject, 'borderTopWidth' | 'borderTopStyle' | 'borderTopColor'>;
+
+export function borderTop(width: BorderWidthProperty<MakeStylesCSSValue>): BorderTopStyle;
+export function borderTop(width: BorderWidthProperty<MakeStylesCSSValue>, style: BorderStyleProperty): BorderTopStyle;
 export function borderTop(
   width: BorderWidthProperty<MakeStylesCSSValue>,
   style: BorderStyleProperty,
   color: BorderColorProperty,
-): MakeStylesStrictCSSObject;
+): BorderTopStyle;
 
 /**
  * A function that implements expansion for "border-top", it's simplified - check usage examples.
@@ -24,10 +23,10 @@ export function borderTop(
  */
 export function borderTop(
   ...values: [BorderWidthProperty<MakeStylesCSSValue>, BorderStyleProperty?, BorderColorProperty?]
-): MakeStylesStrictCSSObject {
+): BorderTopStyle {
   return {
     borderTopWidth: values[0],
-    ...(values[1] && ({ borderTopStyle: values[1] } as MakeStylesStrictCSSObject)),
+    ...(values[1] && ({ borderTopStyle: values[1] } as BorderTopStyle)),
     ...(values[2] && { borderTopColor: values[2] }),
   };
 }