+++ packages/cra-template/CHANGELOG.md
@@ -1,8 +1,53 @@
 # Change Log - @fluentui/cra-template
 
-This log was last generated on Wed, 08 Dec 2021 07:28:19 GMT and should not be manually modified.
+This log was last generated on Thu, 16 Dec 2021 07:26:39 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [8.0.48](https://github.com/microsoft/fluentui/tree/@fluentui/cra-template_v8.0.48)
+
+Thu, 16 Dec 2021 07:26:39 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/cra-template_v8.0.47..@fluentui/cra-template_v8.0.48)
+
+### Patches
+
+- Bump @fluentui/react to v8.48.0 ([PR #21011](https://github.com/microsoft/fluentui/pull/21011) by beachball)
+
+## [8.0.47](https://github.com/microsoft/fluentui/tree/@fluentui/cra-template_v8.0.47)
+
+Wed, 15 Dec 2021 07:31:29 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/cra-template_v8.0.46..@fluentui/cra-template_v8.0.47)
+
+### Patches
+
+- Bump @fluentui/react to v8.47.3 ([PR #20716](https://github.com/microsoft/fluentui/pull/20716) by beachball)
+
+## [8.0.46](https://github.com/microsoft/fluentui/tree/@fluentui/cra-template_v8.0.46)
+
+Tue, 14 Dec 2021 07:30:02 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/cra-template_v8.0.45..@fluentui/cra-template_v8.0.46)
+
+### Patches
+
+- Bump @fluentui/react to v8.47.2 ([PR #20904](https://github.com/microsoft/fluentui/pull/20904) by beachball)
+
+## [8.0.45](https://github.com/microsoft/fluentui/tree/@fluentui/cra-template_v8.0.45)
+
+Fri, 10 Dec 2021 07:30:35 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/cra-template_v8.0.44..@fluentui/cra-template_v8.0.45)
+
+### Patches
+
+- Bump @fluentui/react to v8.47.1 ([PR #20587](https://github.com/microsoft/fluentui/pull/20587) by beachball)
+
+## [8.0.44](https://github.com/microsoft/fluentui/tree/@fluentui/cra-template_v8.0.44)
+
+Thu, 09 Dec 2021 07:27:18 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/cra-template_v8.0.43..@fluentui/cra-template_v8.0.44)
+
+### Patches
+
+- Bump @fluentui/react to v8.47.0 ([PR #20161](https://github.com/microsoft/fluentui/pull/20161) by beachball)
 
 ## [8.0.43](https://github.com/microsoft/fluentui/tree/@fluentui/cra-template_v8.0.43)